+++ lib/animated_snack_bar.dart
@@ -2,3 +2,4 @@ library animated_snack_bar;
 
 export 'src/animated_snack_bar.dart';
 export 'src/types.dart';
+export 'src/settings/raw_animated_snackbar_position_settings.dart';
