+++ lib/src/animated_snack_bar.dart
@@ -1,3 +1,4 @@
+import 'package:animated_snack_bar/src/settings/raw_animated_snackbar_position_settings.dart';
 import 'package:flutter/material.dart';
 
 import 'package:animated_snack_bar/src/types.dart';
