+++ lib/src/settings/raw_animated_snackbar_position_settings.dart
@@ -1,0 +1,17 @@
+class RawAnimatedSnackBarPositionSettings {
+  final double? topVisible;
+  final double? topInvisible;
+  final double? left;
+  final double? right;
+  final double? bottomVisible;
+  final double? bottomInvisible;
+
+  const RawAnimatedSnackBarPositionSettings({
+    this.topVisible,
+    this.topInvisible,
+    this.left,
+    this.right,
+    this.bottomVisible,
+    this.bottomInvisible,
+  });
+}
