+++ lib/src/widgets/raw_animated_snack_bar.dart
@@ -4,25 +4,11 @@ import 'package:flutter/material.dart';
 
 import 'package:animated_snack_bar/animated_snack_bar.dart';
 
+import '../settings/raw_animated_snackbar_position_settings.dart';
+
 Duration _opacityDuration = const Duration(milliseconds: 400);
 
-class RawAnimatedSnackBarPositionSettings {
-  final double? topVisible;
-  final double? topInvisible;
-  final double? left;
-  final double? right;
-  final double? bottomVisible;
-  final double? bottomInvisible;
 
-  const RawAnimatedSnackBarPositionSettings({
-    this.topVisible,
-    this.topInvisible,
-    this.left,
-    this.right,
-    this.bottomVisible,
-    this.bottomInvisible,
-  });
-}
 
 class RawAnimatedSnackBar extends StatefulWidget {
   const RawAnimatedSnackBar({
